+++ temba/orgs/models.py
@@ -43,7 +43,6 @@ from temba.utils.currencies import currency_for_country
 from temba.utils.email import send_template_email, send_simple_email, send_custom_smtp_email
 from temba.utils.models import SquashableModel
 from temba.utils.text import random_string
-from temba.utils.timezones import timezone_to_country_code
 from timezone_field import TimeZoneField
 from urlparse import urlparse
 from uuid import uuid4
@@ -290,6 +289,9 @@ class Org(SmartModel):
     def get_branding(self):
         from temba.middleware import BrandingMiddleware
         return BrandingMiddleware.get_branding_for_host(self.brand)
+
+    def get_brand_domain(self):
+        return self.get_branding()['domain']
 
     def lock_on(self, lock, qualifier=None):
         """
@@ -788,13 +790,13 @@ class Org(SmartModel):
         nexmo_uuid = str(uuid4())
         nexmo_config = {NEXMO_KEY: api_key.strip(), NEXMO_SECRET: api_secret.strip(), NEXMO_UUID: nexmo_uuid}
         client = NexmoClient(key=nexmo_config[NEXMO_KEY], secret=nexmo_config[NEXMO_SECRET])
-        temba_host = settings.TEMBA_HOST.lower()
+        domain = self.get_brand_domain()
 
-        app_name = "%s/%s" % (temba_host, nexmo_uuid)
+        app_name = "%s/%s" % (domain, nexmo_uuid)
 
-        answer_url = "https://%s%s" % (temba_host, reverse('handlers.nexmo_call_handler', args=['answer', nexmo_uuid]))
+        answer_url = "https://%s%s" % (domain, reverse('handlers.nexmo_call_handler', args=['answer', nexmo_uuid]))
 
-        event_url = "https://%s%s" % (temba_host, reverse('handlers.nexmo_call_handler', args=['event', nexmo_uuid]))
+        event_url = "https://%s%s" % (domain, reverse('handlers.nexmo_call_handler', args=['event', nexmo_uuid]))
 
         params = dict(name=app_name, type='voice', answer_url=answer_url, answer_method='POST',
                       event_url=event_url, event_method='POST')
@@ -1792,40 +1794,6 @@ class Org(SmartModel):
             add_component(component)
 
         return all_components
-
-    def get_recommended_channel(self):
-        from temba.channels.views import TWILIO_SEARCH_COUNTRIES
-        NEXMO_RECOMMEND_COUNTRIES = ['US', 'CA', 'GB', 'AU', 'AT', 'FI', 'DE', 'HK', 'HU',
-                                     'LT', 'NL', 'NO', 'PL', 'SE', 'CH', 'BE', 'ES', 'ZA']
-
-        countrycode = timezone_to_country_code(self.timezone)
-        recommended = 'A'
-
-        if countrycode in [country[0] for country in TWILIO_SEARCH_COUNTRIES]:
-            recommended = 'T'
-
-        elif countrycode in NEXMO_RECOMMEND_COUNTRIES:
-            recommended = 'NX'
-
-        elif countrycode == 'KE':
-            recommended = 'AT'
-
-        elif countrycode == 'ID':
-            recommended = 'H9'
-
-        elif countrycode == 'SO':
-            recommended = 'SQ'
-
-        elif countrycode == 'NP':  # pragma: needs cover
-            recommended = 'BM'
-
-        elif countrycode == 'UG':  # pragma: needs cover
-            recommended = 'YO'
-
-        elif countrycode == 'PH':  # pragma: needs cover
-            recommended = 'GL'
-
-        return recommended
 
     def increment_unread_msg_count(self, type):
         """