+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+v3.0.258
+----------
+ * Set brand domain on channel creations, use for callbacks
+
+v3.0.257
+----------
+ * Update path backfill migration to first migrate all flows to 10.4 which fixes missing exit_uuids and action UUIDs
+ * Ensure actions have UUIDs in single message and join-group flows
+ * Flow migration command shouldn't blow up if a single flow fails
+
+v3.0.256
+----------
+ * Migration to populate run paths (timeconsuming, may want to fake aand run manually)
+
+v3.0.255
+----------
+ * Fix Twilio to redirect to twilio claim page after connecting Twilio
+ * Add FlowRun.path and start populating it for new flow steps
+ * Removes no longer used Msg.has_template_error field
+
+v3.0.254
+----------
+ * Use get_host() when calculating signature for voice callbacks
+
+v3.0.253
+----------
+ * use get_host() when validating IVR requests
+
+v3.0.252
+----------
+ * Better Twilio channel claiming
+
+v3.0.250
+----------
+ * Tweaks to recommended channels display
+
 v3.0.246
 ----------
  * Update smartmin to version 1.11.4